Added deeper detections and ability to use starter kits in with Vagrant

+++ Vagrantfile
@@ -18,14 +18,17 @@ Vagrant.configure("2") do |config|
   # within the machine from a port on the host machine and only allow access
   # via 127.0.0.1 to disable public access
   
-  config.vm.network "forwarded_port", guest: 80, host: 8001, host_ip: "127.0.0.1"
+  config.vm.network "forwarded_port", guest: 80, host: 8001, host_ip: "127.0.0.1", auto_correct: true
 
   # Share an additional folder to the guest VM. The first argument is
   # the path on the host to the actual folder. The second argument is
   # the path on the guest to mount the folder. And the optional third
   # argument is a set of non-required options.
 
-  # config.vm.synced_folder "../data", "/vagrant_data"
+  config.vm.synced_folder "dist", "/var/www", 
+    create: true,
+    owner: "www-data", 
+    group: "www-data"
 
   # Enable provisioning with a shell script. Additional provisioners such as
   # Ansible, Chef, Docker, Puppet and Salt are also available. Please see the
@@ -39,11 +42,6 @@ Vagrant.configure("2") do |config|
   config.vm.provision "file", source: "config/apache/000-default.conf", destination: "000-default.conf"
 
   config.vm.provision "shell", inline: <<-SHELL
-
-    #Add environment variables
-    
-    export $(egrep -v '^#' .env | xargs)
-    rm .env
 
     # Suppress interactive frontend.
 
@@ -60,6 +58,7 @@ Vagrant.configure("2") do |config|
       openssl \
       imagemagick \
       expect \
+      dos2unix \
       php \
       php-cli \
       php-common \
@@ -70,6 +69,13 @@ Vagrant.configure("2") do |config|
       php-xml \
       php-gd 
 
+    #Add environment variables
+
+    dos2unix .env
+    export $(egrep -v '^#' .env | xargs)
+    rm .env
+    export COMPOSER_ALLOW_SUPERUSER=1
+
     # Add our Apache configurations.
 
     mv -f apache2.conf /etc/apache2/apache2.conf
@@ -77,17 +83,22 @@ Vagrant.configure("2") do |config|
 
     # Remove the DocumentRoot only if Laravel is not detected. We'll be putting it back together later.
 
-    if [ ! -f /var/www/artisan ]; then
+    if [ ! -f /var/www/composer.json ]; then
       rm -rf /var/www/*
     fi
 
     # Install Composer
 
-    curl -sS https://getcomposer.org/installer -o composer-setup.php
-    php composer-setup.php --install-dir=/usr/local/bin --filename=composer
-    rm composer-setup.php
-    export COMPOSER_ALLOW_SUPERUSER=1
-    composer config -g github-oauth.github.com $GITHUB_AUTH_TOKEN
+    if ! composer > /dev/null; then
+      echo "Composer not installed. Installing."
+      curl -sS https://getcomposer.org/installer -o composer-setup.php
+      php composer-setup.php --install-dir=/usr/local/bin --filename=composer
+      rm composer-setup.php
+      composer config -g github-oauth.github.com $GITHUB_AUTH_TOKEN
+    else
+      echo "Composer is already installed. Continuing..."
+    fi
+
 
     # Make a swap file to avoid Composer failing.
     # https://getcomposer.org/doc/articles/troubleshooting.md#proc-open-fork-failed-errors
@@ -106,9 +117,9 @@ Vagrant.configure("2") do |config|
 
     fi
 
-    # Statamic is a go.
+    # Install Statamic if the composer.json isn't found.
 
-    if [ ! -d "/var/www/html" ]; then
+    if [ ! -f /var/www/composer.json ]; then
 
       # We previously removed the DocumentRoot above. This installs Statamic and recreates the DocumentRoot.
 
@@ -123,14 +134,33 @@ Vagrant.configure("2") do |config|
 
     else
 
-      # If the DocumentRoot exists, we know we've already installed Statamic.
+      # If the composer.json exists, we know we've already installed Statamic.
 
-      echo "DocumentRoot exists. Statamic is already installed."
+      if [ ! -d "/var/www/vendor" ]; then
+        echo "Existing Statamic website detected, but vendor folder is not installed."
+        composer install -d /var/www/
+        cp /var/www/.env.example /var/www/.env && php /var/www/artisan key:generate
+      fi
 
     fi
 
-    expect statamic.exp
-    sleep 6
+    if ! [[ -z "${STATAMIC_ADMIN_EMAIL}" ]] && ! [[ -z "${STATAMIC_ADMIN_PASS}" ]]; then
+
+      users=`ls -1 /var/www/users/*.yaml 2>/dev/null | wc -l`
+      if [ $users == 0 ]; then
+        echo "No Statamic users detected. Creating initial user."
+        echo "Creating initial Statamic user."
+        expect statamic.exp
+        sleep 6
+      else
+        echo "Existing Statamic users detected. Skipping user creation."
+      fi
+    
+    else
+
+      echo "No Statamic credentials found. Continuing..."
+
+    fi
 
     # Hand everything back to Apache and enable mod_rewrite.
     a2enmod rewrite
@@ -141,8 +171,4 @@ Vagrant.configure("2") do |config|
 
   SHELL
 
-  config.vm.synced_folder ".www", "/var/www", 
-    create: true,
-    owner: "www-data", 
-    group: "www-data"
 end
